+++ app/Models/Contact/Debt.php
@@ -2,9 +2,13 @@
 
 namespace App\Models\Contact;
 
+use App\Helpers\MoneyHelper;
 use App\Models\Account\Account;
 use Illuminate\Database\Eloquent\Builder;
 use App\Models\ModelBindingHasherWithContact as Model;
+use App\Models\Settings\Currency;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Illuminate\Support\Facades\Auth;
 
 /**
  * @property Account $account
@@ -33,6 +37,8 @@ class Debt extends Model
 
     /**
      * Get the account record associated with the debt.
+     *
+     * @return BelongsTo
      */
     public function account()
     {
@@ -41,10 +47,22 @@ class Debt extends Model
 
     /**
      * Get the contact record associated with the debt.
+     *
+     * @return BelongsTo
      */
     public function contact()
     {
         return $this->belongsTo(Contact::class);
+    }
+
+    /**
+     * Get the currency record associated with the debt.
+     *
+     * @return BelongsTo
+     */
+    public function currency()
+    {
+        return $this->belongsTo(Currency::class);
     }
 
     /**
@@ -78,5 +96,53 @@ class Debt extends Model
     public function scopeOwed(Builder $query)
     {
         return $query->where('in_debt', 'no');
+    }
+
+    /**
+     * Set exchange value.
+     *
+     * @return string
+     */
+    public function setAmountAttribute($value)
+    {
+        $currency = $this->currency()->first();
+        if (! $currency) {
+            $currency = Auth::user()->currency;
+        }
+        $this->attributes['amount'] =  MoneyHelper::formatInput($value, $currency);
+    }
+
+    /**
+     * Get value of amount (without currency).
+     *
+     * @return string
+     */
+    public function getAmountAttribute(): string
+    {
+        if (! $this->attributes['amount']) {
+            return '';
+        }
+        $currency = $this->currency()->first();
+        if (! $currency) {
+            $currency = Auth::user()->currency;
+        }
+        return MoneyHelper::exchangeValue($this->attributes['amount'], $currency);
+    }
+
+    /**
+     * Get display value: amount with currency.
+     *
+     * @return string
+     */
+    public function getDisplayValueAttribute(): string
+    {
+        if (! $this->attributes['amount']) {
+            return '';
+        }
+        $currency = $this->currency()->first();
+        if (! $currency) {
+            $currency = Auth::user()->currency;
+        }
+        return MoneyHelper::format($this->attributes['amount'], $currency);
     }
 }